Refactor directory
