+++ netlify.toml
@@ -1,0 +1,10 @@
+[build]
+  functions = "src"
+
+[[redirects]]
+  from = "/*"
+  to = "/.netlify/functions/api"
+  status = 200
+
+[dev]
+  publish = "dist"
