+++ package.json
@@ -16,6 +16,7 @@
     "express": "^4.19.2",
     "helmet": "^7.1.0",
     "mongodb": "^6.6.1",
-    "nodemon": "^3.1.0"
+    "nodemon": "^3.1.0",
+    "serverless-http": "^3.2.0"
   }
 }
